+++ Makefile
@@ -1,9 +1,12 @@
+build: client server
+client:
+	cd cmd/client && go build -o ../../bin/fyle-client
+server:
+	cd cmd/server && go build -o ../../bin/fyle-server
+
 run-server:
-	@cd server && go run main.go
+	@cd cmd/server && go run main.go
 run-client: 
-	@cd client && go run main.go
-test:
-	cd server && go test ./...
-	cd client && go test ./...
+	@cd cmd/client && go run main.go
 
 .PHONY: client server
