--- client/go.mod
@@ -1,3 +1,0 @@
-module github.com/fmich7/fyle-cli
-
-go 1.23.3
