+++ cmd/client/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	"github.com/fmich7/fyle/internal/cli"
+)
+
+func main() {
+	cli.Execute()
+}
