+++ cmd/server/main.go
@@ -4,13 +4,13 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/fmich7/fyle/api"
-	"github.com/fmich7/fyle/storage"
+	"github.com/fmich7/fyle/pkg/api"
+	"github.com/fmich7/fyle/pkg/storage"
 )
 
 func main() {
 	PORT := ":3000"
-	store := storage.NewDiskStorage("./uploads")
+	store := storage.NewDiskStorage("../../uploads")
 
 	server := api.NewServer(PORT, store)
