+++ go.mod
@@ -1,0 +1,9 @@
+module github.com/fmich7/fyle
+
+go 1.23.3
+
+require (
+	github.com/inconshreveable/mousetrap v1.1.0 // indirect
+	github.com/spf13/cobra v1.8.1 // indirect
+	github.com/spf13/pflag v1.0.5 // indirect
+)
