+++ internal/cli/root.go
@@ -1,0 +1,22 @@
+package cli
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/cobra"
+)
+
+var rootCmd = &cobra.Command{
+	Use:   "fyle",
+	Short: "fyle is a cli tool for managing your files on the cloud",
+	Run: func(cmd *cobra.Command, args []string) {
+
+	},
+}
+
+func Execute() {
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Fprintf(os.Stderr, "Error on executing cli: %s\n", err)
+	}
+}
