+++ internal/cli/upload.go
@@ -1,0 +1,5 @@
+package cli
+
+func UploadFile() {
+
+}
