+++ pkg/storage/storage.go
@@ -1,6 +1,6 @@
 package storage
 
-import "github.com/fmich7/fyle/types"
+import "github.com/fmich7/fyle/internal/types"
 
 type Storage interface {
 	UploadFile(file *types.File) error
